Updates file to try new module interface
