+++ SDK/Tests/ClientTests/Client+FhirStu3Tests/Data4LifeClient+FhirStu3Tests.swift
@@ -33,7 +33,7 @@ class Data4LifeClientFhirStu3Tests: XCTestCase {
     var appDataService: AppDataServiceMock!
     var keychainService: KeychainServiceMock!
     var recordService: RecordServiceMock<DocumentReference,DecryptedFhirStu3Record<DocumentReference>>!
-    var environment: Environment!
+    var environment: D4LEnvironment!
     var versionValidator: SDKVersionValidatorMock!
 
     override func setUp() {
